# api/openapi.yaml
openapi: 3.0.3
info:
  title: Morning Night Dream - AppGateway
  description: |-
    This is the AppGateway API documentation.
  termsOfService: https://example.com
  contact:
    email: morning.night.dream@example.com
  license:
    name: MIT
  version: 0.0.1
externalDocs:
  description: Find out more about Swagger
  url: http://swagger.io
servers:
  - url: http://localhost:8082/api
tags:
  - name: auth
    description: 認証
  - name: article
    description: 記事
  - name: health
    description: ヘルスチェック
  - name: version
    description: バージョン
paths:
  /v1/auth/signup:
    post:
      tags:
        - auth
      summary: サインアップ
      description: サインアップ
      operationId: v1AuthSignUp
      security:
        - apiKey: []
      requestBody:
        required: true
        description: サインアップリクエストボディ
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: メールアドレス
                  format: email
                  example: morning.night.dream@example.com
                password:
                  type: string
                  description: パスワード
                  format: password
                  example: password
              required:
                - email
                - password
      responses:
        '200':
          description: 成功
        '400':
          description: Bad Request
  /v1/auth/signin:
    post:
      tags:
        - auth
      summary: サインイン
      description: サインイン
      operationId: v1AuthSignIn
      requestBody:
        required: true
        description: サインインリクエストボディ
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  description: メールアドレス
                  format: email
                  example: morning.night.dream@example.com
                password:
                  type: string
                  description: パスワード
                  format: password
                  example: password
                publicKey:
                  type: string
                  description: 公開鍵
                  format: base64
                expiresIn:
                  type: integer
                  description: トークン有効期限(秒)
                  default: 86400
                  minimum: 1
                  maximum: 86400
              required:
                - email
                - password
                - publicKey
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
  /v1/auth/verify:
    get:
      tags:
        - auth
      summary: 検証
      description: 検証
      operationId: v1AuthVerify
      security:
        - idToken: []
        - sessionToken: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1UnauthorizedResponse'
        '500':
          description: Internal Server Error
  /v1/auth/refresh:
    get:
      tags:
        - auth
      summary: リフレッシュ
      description: リフレッシュ
      operationId: v1AuthRefresh
      security:
        - sessionToken: []
      parameters:
        - name: code
          in: query
          description: 署名付きコード
          required: true
          explode: true
          schema:
            type: string
        - name: signature
          in: query
          description: 署名
          required: true
          explode: true
          schema:
            type: string
        - name: expiresIn
          in: query
          required: false
          explode: true
          schema:
            type: integer
            description: トークン有効期限(秒)
            default: 86400
            minimum: 1
            maximum: 86400
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
  /v1/auth/signout:
    get:
      tags:
        - auth
      summary: サインアウト
      description: サインアウト
      operationId: v1AuthSignOut
      security:
        - idToken: []
        - sessionToken: []
      responses:
        '200':
          description: OK
  /v1/auth:
    delete:
      tags:
        - auth
      summary: リサイン(退会)
      description: リサイン(退会)
      operationId: v1AuthResign
      security:
        - idToken: []
        - sessionToken: []
      requestBody:
        required: true
        description: リサインリクエストボディ
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: パスワード
                  format: password
                  example: password
              required:
                - password
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
  /v1/sign:
      get:
        deprecated: true
        tags:
          - auth
        summary: 署名検証
        description: 署名検証
        operationId: v1Sign
        parameters:
          - name: code
            in: query
            description: 署名付きコード
            required: true
            explode: true
            schema:
              type: string
          - name: signature
            in: query
            description: 署名
            required: true
            explode: true
            schema:
              type: string
        responses:
          '200':
            description: OK
          '401':
            description: Unauthorized
  /v1/article:
    get:
      tags:
        - article
      summary: List articles
      description: List articles
      operationId: v1ListArticles
      parameters:
        - name: pageToken
          in: query
          description: トークン
          required: false
          explode: true
          schema:
            type: string
        - name: maxPageSize
          in: query
          description: ページサイズ
          required: true
          explode: true
          schema:
            type: integer
            minimum: 1
            maximum: 20
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/V1ListArticleResponse'
        500:
          description: サーバーエラー
  /v1/health:
    get:
      tags:
        - health
      summary: ヘルスチェック
      description: ヘルスチェック
      operationId: v1Health
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error
  /v1/version/api:
    get:
      tags:
        - version
      summary: APIバージョン
      description: APIバージョン
      operationId: v1APIVersion
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error
  /v1/version/core:
    get:
      tags:
        - version
      summary: Coreバージョン
      description: Coreバージョン
      operationId: v1CoreVersion
      responses:
        200:
          description: OK
        500:
          description: Internal Server Error
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: api-key
      in: header
    idToken:
      type: apiKey
      name: id_token
      in: cookie
    accessToken:
      type: apiKey
      name: access_token
      in: cookie
    sessionToken:
      type: apiKey
      name: session_token
      in: cookie
  schemas:
    Article:
      type: object
      properties:
        id:
          type: string
          description: id
          format: uuid
        url:
          type: string
          description: 記事のURL
          format: uri
          example: https://example.com
        title:
          type: string
          description: タイトル
          example: sample title
        description:
          type: string
          description: description
          example: sample description
        thumbnail:
          type: string
          description: サムネイルのURL
          format: uri
          example: https://example.com
        tags:
          type: array
          description: タグ
          items:
            type: string
            example: tag
    V1ListArticleResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: '#/components/schemas/Article'
        nextPageToken:
          type: string
          description: 次回リクエスト時に指定するページトークン
    V1UnauthorizedResponse:
      type: object
      properties:
        code:
          type: string
          description: コード
          format: uuid
      required:
        - code